make_announcement: adding admin checks

+++ app/views/do.py
@@ -802,9 +802,11 @@ def deleteannouncement():
 
 
 @do.route("/do/makeannouncement", methods=['POST'])
-@login_required
 def make_announcement():
     """ Flagging post as announcement - not api """
+    if not current_user.is_admin():
+        abort(404)
+
     form = DeletePost()
 
     if form.validate():
